Refactor `control_flow.py` into several smaller pieces
